small correctness fix
This seems to resolve the sporadic memory issues, although it shouldn't, reasonably. The (lack of) documentation for GetIconRefFromFile() doesn't specify any guarantees about results on error conditions, but getting a ref even with an error seems like an oversight, at best.

Also, to note, we should likely be releasing the last reference that we still hold before exiting the program. Although /our/ memory is cleaned up on exit, it sounds as if this may not be ours, and we'd be preventing the release of one icon per run if so. -- That will touch a lot of possible exit points though; so i'm leaving it for a later commit.